Add FillType to more classes

+++ gsSlicer/gsSlicer/fill/basic/BasicFillCurve.cs
@@ -1,4 +1,5 @@
 ﻿using g3;
+using System;
 using System.Collections.Generic;
 
 namespace gs
@@ -18,9 +19,17 @@ namespace gs
                 BeginOrAppendCurve(vertexEnumerator.Current);
         }
 
-        public void Trim(double v)
+        [Obsolete("This method currently is jury-rigged; logic needs to be improved for non-trivial cases")]
+        public void TrimBothEnds(double v)
         {
+            // TODO: This method needs big improvements. The segment and vertex information needs
+            // to be appropriately interpolated when removing points.
+
             Polyline.Trim(v);
+
+            // Exit early if no points were actually removed
+            if (Polyline.VertexCount == VertexInfo.Count)
+                return;
 
             // Remove any vertex info that was trimmed away.
             VertexInfo.RemoveRange(Polyline.VertexCount, VertexInfo.Count - Polyline.VertexCount);
